From: Nathaniel Plant [mailto:nplant@usgs.gov] Sent: Monday, July 09, 2012 9:30 AM To: 'Aaron Turecek' Cc: 'rthieler@usgs.gov'; 'Ben Gutierrez' Subject: RE: Netica stuff
Hi Aaron,
I’m about to send you threads to the scenario viz project that we started with John Aguinaldo.  I think that this is what we are after.  It will have some cross-over with other tasks and the example I’m giving you is out of date. But, it is nicely structured.

Step 1. Get the input  data to drive a Bayesian prediction. In this case, read in the CVI data and write out a bunch of case files that play with modifications to SLR and uncertainty. This determines which scenarios would be available for inspection.  It also lays out a mechanism for creating new scenarios on the fly, should we want to do that.  

Files are the “convert…Shapefiles3.m”, matlab that does the work
A data set “data0/cvi_v01_733506.dat”, where data0/ indicates folder for actual data as opposed to mock data that will be used in scenarios. 
A net “BayesCVI_v2em_loaded.neta”, old version but useful to make sure the machinery works before updating everything