From: Brown, Jenna [mailto:jennabrown@usgs.gov] Sent: Thursday, June 04, 2015 10:29 AM To: Nathaniel Plant Subject: m-file to compile, please :)
When you get a chance, can you please compile this code for me?

Basically, the code connects to the CIL ftp server and compares the timestamps of the remote and local directories for a given date and updates our local archive if any images/products have been updated or were missing. 

I wrote a batch file that I will run daily to figure out what the date was a week ago and then check/update that folder (so any major differences in the remote and local directories are being updated one day at a time). I found most folders were last updated 6 days after that date, so this should keep things up to date.

I also wrote a batch file to run the "check/update" code for the past 60 days, so I'll run it every month as a double-check.

Let me know if you have any thoughts or suggestions to improve our Argus backup system.

Thanks!!!!