Fixing an R2015a incompatability as follows:
Re: Code made with Guide broken by graphics changes    [ ref:_00Di0Ha1u._500i0RfKVt:ref ]
Hello Marinna, 

I am writing in reference to your Technical Support Case #01397942 regarding 'Code made with Guide broken by graphics changes'. 

Thank you for sending me the necessary files. I was able to run your code and reproduce the error messages. 

After consulting with my colleagues, I was able to identify the cause of the error. After clearing the axes in the "updateplot" function, the line object that is pointed to by "hObject" is deleted. 

In previous versions of MATLAB, the objects were numbered for reference and after clearing the line object, a creation of a new line would create an object with the same handle number as the original. After the graphics update in MATLAB R2014b, the objects are not numbered for reference and "guidata" is being passed a handle to a deleted object. Since "hObject" is pointing to a deleted object, "guidata" throws an error. 

"guidata" will find the parent figure of the first argument, so one work around is to pass the figure handle directly by replacing the following:

 if isobject(hObject),
	guidata(hObject, handles);
 end

with: 

 guidata(handles.figure1,handles)

Please try to run the app and check to see if the error messages are still occurring. Also, please feel free to contact us if you have any further questions or concerns.

Please preserve the Reference ID below in further correspondence on this query. This allows our systems to automatically associate your reply to the appropriate Case. 

If you have a new technical support question, please submit a new request here: 
http://www.mathworks.com/support/servicerequests/create.html

Sincerely,
Aditya Bhatter
MathWorks Technical Support Department